Revert "Revert "Added documentation to new InformationGainFormula class.""
This reverts commit 605727f713cc13a9b40d7dd11b1c6373c404f8f7.

+++ palladian-retrieval/src/test/java/ws/palladian/retrieval/search/web/VimeoSearcherTest.java
@@ -25,7 +25,7 @@ public class VimeoSearcherTest {
         assertEquals("https://vimeo.com/20800127", result1.getUrl());
         assertEquals(246, (long)result1.getRunTime());
         
-        long resultCount = VimeoSearcher.parseResultCount(jsonString);
+        int resultCount = VimeoSearcher.parseResultCount(jsonString);
         assertEquals(94609, resultCount);
     }
